Add support for setting CNI options
Add a new field to the cluster config, `CniOptions`, that allows the
user to configure options for the Amazon VPC CNI plugin for Kubernetes.

This is currently implemented by shelling out to kubectl from a dynamic
resource. Once we are able to programatically take ownership of existing
resources, we should use proper k8s resources instead.