Filter public subnets as necessary (#21)
If both public and private subnets are attached to an EKS cluster, the
API server will only be exposed to the private subnets. Any workers
attached to the public subnets will be unable to contact the API server
and will fail to register as nodes. These changes filter public subnets
from the set of subnets passed to the worker nodes iff the EKS cluster
is also attached to private subnets.

Fixes #18.