Allow explicit node subnets
And improve subnet filtering. Explicit subnets for worker nodes can be
specified if the automatic filtering does not work as expected.
The automatic filtering has been improved by using the presence of a
route to an internet gateway in a subnet's route table as the indicator
that the subnet is a public subnet.