fix(cni): modify CNI filepath to store the relative path only in state
This fix gets around the absolute path being inserted into the state for
the dynamic provider, by only storing the relative path for the CNI YAML
file.

The absolute path has now been pushed out to be formed and resolved in
`computeVpcCniYaml()` before the file is opened, instead of the
depending on the full path from the `applyVpcCniYaml()` caller.