test(private-workers): rm as it only tests nodeAssociatePublicIpAddress
This test is being removed as it is extraneous, and removing it would
also be a quality of life improvement on CI runs as it's taking about ~
1 hour each these days.

The test was setting `args.nodeAssociatePublicIpAddress` which gets
passed into `aws.ec2.LaunchConfiguration`, but it wasn't verifying the
setting. Plus, relying on `pulumi-aws` to ensure it works as advertised
seemed appropriate.

https://github.com/pulumi/pulumi-eks/blob/5d410705fb8146ba37ab2445fd44a13f13689a40/nodejs/eks/nodegroup.ts#L427-L449