refactor(vpcCni): set node anti-affinity to not deploy to fargate
This change sets a node anti-affinity for the AWS CNI DaemonSet to only
deploy to nodes that are not labeled for fargate with
`eks.amazonaws.com/compute-type=fargate`.

By default, the DaemonSet uses a `RollingUpdate`
[strategy](https://kubernetes.io/docs/tasks/manage-daemon/update-daemon-set/#performing-a-rolling-update)
with a `maxUnavailable` of 1 pod. This means that on a Pulumi update to
this change for existing clusters, the `aws-node` pods will update one
at a time, until all are updated with the new node anti-affinity. This
will not cause down time of the cluster or running pods.

In summary, this change allows clusters to:

- Mix and match the type of nodes used (EC2 / self-managed, AWS managed,
and/or Fargate)
- Deploy the aws-cni to the cluster for any configuration of nodes, and
- Although the DS will always be created, it will only deploy to nodes
that are not fargate backed.