Extend Cluster Timeouts
Fixes: #340

Sometimes clusters can time out whne creating, to allow this not
to be the case, we are going to specify a create timeout of 1hour
by default. This means the user *should not* need to worry about
custom timeouts.

For good measure, I added a delete timeout extension as well since
that can sometimes that a long time as well