test(replace-secgroup): use cluster as parent
Using the awsx vpc as the parent for the sg in 2.0 returns the following
error, so we move away to a ComponentResource such as the cluster for
the parent.

```
index.ts(31,4): error TS2345: Argument of type 'Vpc' is not
assignable to parameter of type 'ComponentResource<any>'.
Property 'initialize' is protected but type 'Vpc' is not a
class derived from 'ComponentResource<TData>'.
```