feat(cluster): add encryptionConfigKeyArn opt to encrypt k8s Secrets
EKS supports envelope encryption for Kubernetes Secrets that use an AWS
KMS key to sign the Secrets in etcd.

A new cluster option `encryptionConfigKeyArn` is being added that allows
the configuration of a KMS key ARN to be provided for establishing the
cluster's encryption configuration.

Note: the encryption configuration becomes a core part of the cluster's
configuration. Any changes to its settings will require rebuilding a new cluster
and nodes.

See for more details.
https://aws.amazon.com/blogs/containers/using-eks-encryption-provider-support-for-defense-in-depth/