fix: Fetch partition once add explict provider
I wrote the function to return the string `aws` if the result of
grabbing the partition is false. I believe the only dependency to
running the parition function is having aws:region set in the config
yaml of the call. Could be worthwhile to check, but considering the
current operation of the module its safe to assume that `aws` won't
change anything for those using the module today.