Fetch aws partition (#788)
* feat: Dyamically fetch aws partition

Initial code to dynamically fetch the partition and interpolate that in
place of the hardcoded IAM arns. This will allow gov cloud, china, and
other regions that use different partitions to provision EKS clusters
without having to use a transformation workaround which can be hard to
implement in some languages like Golang

Related to #570,#534,#386

* fix: Fetch partition once add explict provider

I wrote the function to return the string `aws` if the result of
grabbing the partition is false. I believe the only dependency to
running the parition function is having aws:region set in the config
yaml of the call. Could be worthwhile to check, but considering the
current operation of the module its safe to assume that `aws` won't
change anything for those using the module today.

* fix: Linting errors

Signed-off-by: Anthony Butt <cheeks>

* Use outputs for partitions

- Load once at the beginning of createCore then re-use.
- Allow service role managedPolicyArns to be inputs (to match the underlying resource).
- Use pulumi.interpolate to safely interpolate outputs within a string.

* Separate policy identifier from actual ARN

- Identifier needs to be "plain" (not an input) as it's being used in the resource name and we don't want to wrap the resource creation in an apply so we can retain accurate previews.
- Use the original aws-specific ARN for the identifier to maintain backward compatibility and not change resource names.

* Dynamically set partition for RoleProvider

- Remove left-over doc comment.
- Refactor getRoleProvider to use output interpolation.
- Add to changelog.

Signed-off-by: Anthony Butt <cheeks>
Co-authored-by: Anthony Butt <cheeks>