Remove nodeRootVolumeSize default value from schema
Previously a default for nodeRootVolumeSize was set in the schema (and therefore the SDKs),
which would mean the value was always set. This would cause the mutual exclusion check for
nodeGroupOptions to always fail if nodeGroupOptions was ever supplied. This is not needed,
as the platform itself defaults the value.