Guard against cluster certificate authority being undefined
Likely addresses #676

Although the types don't describe the certificate authority as being possibly undefined, the error reported would indicate it can be. Therefore we'll add some extra guards just in case.

If the cluster certificate authority is undefined, then we'll pass the certData to generateKubeConfig as undefined. This will then not set the property within the kubeconfig, but at least it won't throw a hard error causing the whole deployment to fail.

Aside: Fix resolve() call which requires a value to be passed.