fix: always set AMI ID for custom LaunchTemplates with ManagedNodeGroups (#1166)
### Proposed changes

#1163 enabled setting custom user data in MIME format. However, AWS will
attempt to merge the user data with its own copy if an AMI ID is not
supplied in the LaunchTemplate. This results in the nodes being unable
to join the cluster due to conflicting bootstrap script runs. This PR
explicitly sets the ami id within the launch template.

Testing:

- Updated the existing tests to provision a cluster within a VPC to
surface any node joining issues which CI was previously ignoring

Fixes: #1165