tests: set amiID for upgrade tests (#1182)
### Proposed changes

Explicitly set the AMI ID for our nodes and clusters to prevent
constantly needing to update our test snapshots. The current AMI set in
this PR is obtained from what we currently store in our test snapshots
(example:
https://github.com/pulumi/pulumi-eks/blob/b19f15552956f80a45493f887c3bf32e6cea08d0/provider/testdata/recorded/TestProviderUpgrade/tags/2.5.2/stack.json#L2051).

Explicitly setting the AMI ID is a much lower engineering lift than
attempting to parse the automation API's change summary since it only
returns an operation type/enum, and doesn't include what field is to be
replaced. Furthermore, blanket ignoring a replacement of AMI IDs may
inadvertently mask some replacement bugs we may have in the future.
Note: we also have existing integration tests to also test that the
provider can obtain and use a recommended ID., so we are not losing
coverage of this logic flow.

### Related issues (optional)
Fixes: #1181