Fix ManagedNodeGroup for cluster with API authentication mode
Adding a ManagedNodeGroup to a cluster with API authentication mode fails because
there's verification logic that expects the role for the EC2 instances to be present
in the instanceRoles list of the cluster.
If the necessary authentication configuration for the EC2 instances
was added as access entries, this verification will fail.

This change fixes that by excluding this check in case the cluster supports access entries.