fix: remove ClusterInternal type for AddonOptions
The ClusterInternal type is a internal TS type. When users attempt to
transpile their typescript programs to javascript, they'll encounter a
symbol not found error, since internal types are not exposed in the
type annotations file. ClusterInternal isn't necessary since its used
for MLC codegen, so this will not be a breaking change.