Remove ClusterInternal reference in typescript provider (#1209)
### Proposed changes

The ClusterInternal type is a internal TS type. When users attempt to
transpile their typescript programs to javascript, they'll encounter a
symbol not found error, since internal types won't be exposed in the
type annotations file. `ClusterInternal` isn't necessary since its used
for MLC codegen, so this will not be a breaking change.

#### Testing:

- Updated existing typescript test to also run the build step

### Related issues (optional)

Fixes: #1205