Fix ordering of on-cluster resources (#1221)
The creating of on-cluster resources requires access to the cluster.
Per default, the IAM principal creating the cluster gets admin access.
But the k8s provider that's used to create on-cluster can be configured
to use a different IAM role. If that's done, the auth settings (aws-auth
ConfigMap or access entries) need to be applied before creating the
on-cluster resources.

This change adds the missing dependency links for on-cluster resources.