Allow using user provided version for retrieving node group AMIs
Previously the user provided kubernetes version wasn't used when
looking up the AMI for a node group. Instead the cluster version
was used implicitely.
This fixes that and allows users to specify the kubernetes version.
This also allows users now to upgrade their cluster control plane
and node groups in multiple steps.