Fix ManagedNodeGroup instances not registering with EKS when using enableIMDSv2
When creating a ManagedNodeGroup with enableIMDSv2 set to true and no
kubeletExtraArgs or bootstrapExtraArgs the userdata script is not being created.
This causes the EC2 instances to not register with the EKS cluster.
EKS does not use its default userdata because the Launch Template always specifies
an AMI ID right now. The userdata script is what actually starts the kubelet and
registers the node with the EKS cluster.

This fixes it by only setting the AMI ID if a custom userdata script is provided
in the launch template.