Fix ManagedNodeGroups with custom launch templates always using x86_64 AMI (#1324)
A custom launch template is created for a Node Group when
`kubeletExtraArgs`, `bootstrapExtraArgs` or `enableIMDSv2` is set. In
that case, the recommended AMI is retrieved from SSM Parameter Store.

The input property `instanceType` is used to select the right
architecture. But `ManagedNodeGroup` doesn't have an `instanceType`
input property, it has a list of `instanceTypes`.

This change fixes this by checking the `instanceTypes` property of
Managed Node Groups as well.

Fixes #1323