Support custom userdata for AL2023 (#1332)
This adds the necessary handling for `nodeadm` user data. This is used
for AL2023.

`nodeadm` is a tool used for bootstrapping kubernetes nodes. It's
configuration interface is yaml based and can be set via user data. The
user data needs to be in MIME `multipart/mixed` format. This allows
interleaving the nodeadm configuration with scripts or other user data
entries. See more here:
https://awslabs.github.io/amazon-eks-ami/nodeadm/.