Add support for custom bottlerocket user data (#1334)
### Proposed changes

This change adds the necessary options for configuring settings for
Bottlerocket Operating Systems.
Bottlerocket configuration is driven by a user data script in TOML
format.

The provider will set the base configuration that's necessary for nodes
to successfully register with the kubernetes cluster. Users will have
the ability to add additional settings or override the base
configuration by using the `bottlerocketSettings` parameter. An overview
of the settings can be seen here:
https://bottlerocket.dev/en/os/1.20.x/api/settings/.

### New dependencies
This adds two dependencies to the provider.
1. `@iarna/toml"`: For converting the configuration into valid TOML.
This allows us to expose `bottlerocketSettings` as an object instead of
a string so users do not need to worry about TOML
formatting&serialization
2. `ipaddr.js`: For calculating the `cluster-dns-ip`. We could
definitely create our own IP parser, it's not overly complex for IPv4.
But IPv6 can be more tricky with shortened formats. This is unnecessary
complexity for the provider.

Both of them are maintained libraries with 0 other dependencies.

### Remarks
Bottlerocket will only be supported with `NodeGroupV2` and
`ManagedNodeGroup` components because the older `NodeGroup` (aka
`NodeGroupV1`) uses CloudFormation under the hood and nodes need to
signal that they're ready. Bottlerocket can't execute scripts as part of
the boot up because it doesn't have a shell, so this is not supported.