Support configuring disk size of self-managed bottlerocket node groups (#1337)
The provider wrongly assumed that an AMI only has a single block device.
But Bottlerocket has two.
Bottlerocket has two block devices, the root device stores the OS itself
and the other is for data like images, logs, persistent storage. We need
to allow users to configure the block device for data.

With this change, the provider will choose the block device that gets
modified depending on the OS. If the OS is bottlerocket, the data device
gets modified.
This also adds E2E tests verifying that the node storage capacity
correctly reflects user settings.