Deprecate aws-auth ConfigMap
Following up on AWS' deprecation of the aws-auth ConfigMap.
We're deprecating the enums for the `CONFIG_MAP` and `API_AND_CONFIG_MAP`
AuthenticationModes.
This will steer users into the right direction of using the `API`
AuthenticationMode.

Details around the deprecation are here: https://docs.aws.amazon.com/eks/latest/userguide/auth-configmap.html
This is a good guide for how to migrate away from it: https://docs.aws.amazon.com/eks/latest/userguide/migrating-access-entries.html
We have a Pulumi migration guide for this in docs/authentication-mode-migration.md. As a follow up we want
to publish this to the Pulumi Docs as a How-To.