Align addon configuration (#1370)
Now that the EKS addons are added we need to align them and do some
cleanup. This involves:
- adding the enums introduced in
https://github.com/pulumi/pulumi-eks/pull/1357 to the VPC CNI
- exposing configurationValues for coredns and kube-proxy
- removing kubectl from the provider
- deeply sort addon configuration keys to guarantee stable json
serialization
- remove deepmerge again. It caused issues during unit tests
(https://github.com/voodoocreation/ts-deepmerge/issues/22) and when used
on outputs.

Additionally I discovered and fixed an old bug that luckily never
surfaced. The VPC CNI configuration did incorrectly handle outputs and
called `toString` on them in a couple of places. The increased type
safety and tests around addon configuration uncovered this.

Closes #1369