Update default operating system to AL2023 (#1382)
Driven by the deprecation of AL2 by AWS, we need to ensure users are
deploying node groups with maintained and secure operating systems by
default.
This change adds a `RECOMMENDED` OS enum that points to AL2023 (AWS
default) and uses it as the default for node groups.

The upgrade tests are expected to fail as we're changing defaults. To
re-record we need to first release a new baseline version (e.g. alpha
release)

Closes #1354