Fix output of default node group referring to autoscalingGroupName (#1410)
In https://github.com/pulumi/pulumi-eks/pull/1373 the default node group
was updated to use the `NodeGroupV2` component. We missed changing the
`NodeGroupData` type to reflect this. It was still referring to a
property called `autoScalingGroupName`, but it should've been changed to
expose an `autoScalingGroup`.

Fixes https://github.com/pulumi/pulumi-eks/issues/1402