Fix ManagedNodeGroup taints being wrongly set in userdata (#1441)
The taints for the `ManagedNodeGroup` component were being wrongly
calculated when using custom userdata.
That was the case because the EKS service uses different capitalization
for the taint effect enum than the kubernetes API(e.g. `NO_SCHEDULE` vs
`NoSchedule`).
When building the custom userdata we need to map the EKS style enums to
kubernetes style enums, otherwise it doesn't work.

Fixing this also revealed that taint values being absent aren't
correctly handled either. The change fixes that as well.