Fix ManagedNodeGroup with custom launch template and AMI type
When creating an EKS managed NodeGroup with a custom launch template,
the AMI type must be removed from the nodegroup itself and instead specified
in the launch template.
This did not work if the user specified the AMI type instead of the
provider determining it.

This change fixes that by correctly omitting the user defined
AMI type as well in that case.