Use provider dependencies in schema (#1487)
This PR changes the schema generation to read provider dependency
versions (AWS & Kube) from the `package.json` file (like in
[awsx](https://github.com/pulumi/pulumi-awsx/blob/bb9d7462d8be5279267e2ba651f2de2cb6e54953/schemagen/pkg/gen/schema.go#L264)).
This is the first step towards automating provider dependency upgrades.

To achieve the correct behavior, provider dependencies now need to use a
fixed version instead of a semver range. I used the current versions
form the lock file, except for the aws provider. Versions 6.42 - 6.46 of
the aws provider are not present on mavencentral. To keep the diff of
this PR as small as possible, I upgraded aws to the next possible
version (`6.47.0`).

I'll update all provider versions to latest in a separate PR.