Move integration tests to `./tests` (#1490)
This consolidates integration tests (currently located in `./examples`
and `./provider`) under `./tests`.

This is prep work for https://github.com/pulumi/pulumi-eks/pull/1474,
and the net result is a clearer distinction between "fast" and "slow"
tests meant to be run during the "prerequisites" and "tests" steps in
CI. Everything under `./provider` can be considered "fast" and
everything under `./tests` can be considered "slow" and worthy of
sharding.

Specifically:
* Everything under `./examples/tests` is moved to
`./tests/testdata/programs` as these were assumed to not be user-facing.
* Upgrade tests under `./provider` are also moved to `./tests` but
tagged in a way that they continue to run as their own job. (These are
the new `upgrade_test.go` and `validation_test.go` files.)
* `examples/utils/utils.go` moves to `tests/internal/utils`.
* `TestEksAuthModeUpgrade` wasn't re-enabled as part of #1387, so I
re-enabled and re-recorded it.
* `go.mod` is consolidated at the repo root to keep dependencies
consistent across the provider and tests. (There are a ton of additional
DX advantages to structuring things this way.) This project is not meant
to be consumed as a library so this is an acceptable breaking change.