Add region to kubeconfig (#1498)
The generated kubeconfig is missing the region argument. This can lead
to issues in many scenarios, e.g. when a user has `AWS_DEFAULT_REGION`
set or when they have no default region set and use the provider
configuration for setting the pulumi region (i.e. `aws:region`). This
aligns the provider with how the AWS CLI generates the kubeconfig (`aws
eks update-kubeconfig`).
This change uses the cluster ARN to identify the cluster's region and
sets it in the kubeconfig.

The k8s provider's `kubeconfig` property will not trigger replacements.
Multiple upgrade tests validate this (e.g. `aws-profile`).

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->