Allow creating default instance role even when no default node group is created (#1511)
As part of https://github.com/pulumi/pulumi-eks/issues/1176 we took the
decision to not create the default instance role if users disable the
default node groups. Our assumption back then was that there's not a
substantial need for this role when users opt out of the default node
groups.

https://github.com/pulumi/pulumi-eks/issues/1510 has shown that there is
indeed users that were relying on this default instance role that was
created "by accident". This original change made the upgrade to v3 more
painful than necessary for this set of users!

This change introduces a new plain flag `createInstanceRole` that can be
used to opt-into creating the default instance role even when no default
node group is needed.

Resolves https://github.com/pulumi/pulumi-eks/issues/1510