Clean up end-to-end tests
This removes an obsolete test (managed-ng-os-py), it was added before
nodejs and the rest of the SDKs shared the same implementation. This
functionality is now covered by the nodejs test (managed-ng-os).
Additionally it changes the tests e2e tests to place instances
in private subnets in order to reduce the EIP usage.