Clean up end-to-end tests (#1549)
This removes an obsolete test (managed-ng-os-py), it was added before
nodejs and the rest of the SDKs shared the same implementation. This
functionality is already covered by the nodejs based test
(managed-ng-os).

Additionally it changes the e2e tests to place instances in private
subnets in order to reduce the EIP usage (reduces contention and saves
money).