Fix ManagedNodeGroup ignoring custom AMI if no user data is configured
When users configured a custom AMI for an MNG it only took effect if
their settings also led to custom user data being generated.
This change fixes that by also generating custom user data if users
set a custom AMI. This is necessary because EKS will not provide
default user data if an AMI ID is specified in the launch template.