Fix ManagedNodeGroup ignoring custom AMI if no user data is configured (#1562)
When users specified a custom AMI for an EKS Managed Node Group, the
setting was only taking effect if their configuration already triggered
custom user data generation, otherwise it used the default AMI. To fix
this, I modified the logic to generate custom user data whenever a
custom AMI is configured. This ensures the node group launches
successfully with proper bootstrap settings regardless of other
configuration options (AWS EKS does not provide default user data if an
AMI ID is specified in the launch template).

The previous E2E tests worked because they coincidentally used the same
AMI ID as the default AMI. The updated test intentionally chooses an AMI
for the previous minor version and verifies the EC2 instances are using
this exact AMI.

Fixes #1550