Clean up provider internals (#1591)
This change cleans up the provider internals in preparation for more
involved features like adding EFA support
(https://github.com/pulumi/pulumi-eks/issues/1564). This was necessary
because there was still a substantial amount of legacy cruft in the
provider that wasn't even exposed to users (seem to be leftovers from
moving the nodejs implementation onto MLCs).
Additionally it slightly reorganizes the internal module structure in
order to have a more manageable code structure (compared to the current
flat structure and HUGE files).

This change should be fully transparent to users.