Add EFA support for Managed Node Groups (#1593)
### Proposed changes

This change adds support for configuring
[EFA](https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/efa.html) for
Managed Node Groups. EFA is a high-speed network device for instance to
instance communication, usually used in AI or HPC scenarios.

The change involves adding two new input properties:
- `enableEfaSupport`
- `placementGroupAvailabilityZone`

`placementGroupAvailabilityZone` is necessary because EFA does not work
cross-AZ. We also cannot pick a random AZ based on the subnets because
intersection of subnets and supported AZs (by the instance) is not
stable.
When AWS adds support for certain instance types to new AZs this
selection would change and cause a replacement of the whole node group.

Right now I made `placementGroupAvailabilityZone` a required input when
`enableEfaSupport` is set to true. This way users will not be
susceptible to this foot gun. This is in line with what other tools in
the eco system are doing/planning to do (e.g.
https://github.com/terraform-aws-modules/terraform-aws-eks/blob/50cb230c8a1793f5ef4cc52c4c789b656b141168/modules/eks-managed-node-group/main.tf#L645-L648).

### Related issues (optional)

Closes https://github.com/pulumi/pulumi-eks/issues/1564