Add docs for the ManagedNodeGroup component (#1594)
The ManagedNodeGroup component (like all resources of the eks provider)
are missing docs. As part of adding support for EFA
https://github.com/pulumi/pulumi-eks/issues/1564, I wanted to document
this feature.

This change adds basic docs and a basic example for managed node groups,
as well as an example for EFA.

The examples were created in YAML and then converted to the other
languages.