Increase AWS API retries for test verification (#1600)
Increase the number of retry attempts from 3 to 20 to avoid flakiness.
Many of our test take a similar amount of time to complete and we have
high parallelism, so there's spikes of requests for test verification.
This change will not add substantial time to the test run, but will
reduce flakiness.

Fixes #1598