Fix cron job make entrypoints (#1625)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

When onboarding pulumi-eks to ci-mgmt, our make targets were updated to
be uniform with other providers. This requires us to update our
existing, self-managed, cron job workflow to be updated as well.

A [successful test run of the cron
job](https://github.com/pulumi/pulumi-eks/actions/runs/13082831233)
against this PR was run to validate that this work will unblock CI.

### Related issues (optional)

Fixes: #1605