Bump aws dependency (#1753)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

The AWS provider made a change to the `ec2.SecurityGroup` types making
the `egress` and `ingress` `description` fields non-optional in the
outputs. This caused type-check errors for users who did not have their
AWS dependency pinned. The reason for that is that the EKS provider uses
the `SecurityGroup` output types in one of its input properties.

This PR bumps the AWS SDK version so that these changes are pulled into
EKS. This will make the latest EKS provider compatible with the latest
AWS SDK.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

fixes https://github.com/pulumi/pulumi-eks/issues/1752