Replace tslint with eslint and fix a missing interpolate (#1803)
In https://github.com/pulumi/pulumi-eks/issues/1799 a user reported an
error message surfacing from TS which indicated a missing interpolation.

I wanted to use https://github.com/pulumi/eslint-plugin-pulumi to track
down the source of the problem, but this repo is still using
(deprecated) TSLint.

So I upgraded us to eslint and dropped in the pulumi plugin, which
immediately identified a problem with our invalid subnets error.

Refs https://github.com/pulumi/pulumi-eks/issues/1799.